Script to build ubuntu image for the rpi2
While searching a solution for the Yocto build, here is a simple script to
build rpi2 ubuntu image.

The local.conf file can be modified for customization purpose.

The image is build in three steps: bootstrap, setup and image:

  sudo ./ubuntu-rpi2 bootstrap
  sudo ./ubuntu-rpi2 setup
  sudo ./ubuntu-rpi2 image

Then, the image can be written to the SD card, for instance:

  sudo dd if=vivid/ubuntu-rpi2-vivid-20160516.img of=/dev/sdX bs=10M; sync

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>