Remove the local.inc file
The local.inc contains definitions to the specific build of the end-user. This
file should not be added to the repository. Replace "require" by "include",
then local.inc file is not mandatory to build. If the user wants to override the machine name, then it just needs to copy the local.inc.example file.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>